Move folders out of provider folder.
