+++ src/domains/charts/chartMenuGroup/description.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { Text } from "@netdata/netdata-ui"
-import { withMenuGroup } from "domains/charts/providers/menuGroup"
+import { withMenuGroup } from "domains/charts/menuGroup"
 
 export const ChartMenuGroupDescription = ({ children, ...rest }) =>
   children ? <Text color="border" dangerouslySetInnerHTML={{ __html: children }} {...rest} /> : null
