+++ src/domains/charts/chartMenuGroup/heads.js
@@ -1,6 +1,6 @@
 import React from "react"
-import { ChartHeadsContainer } from "domains/charts/providers/chartHead"
-import { withSubMenu } from "domains/charts/providers/subMenu"
+import { ChartHeadsContainer } from "domains/charts/chartHead"
+import { withSubMenu } from "domains/charts/subMenu"
 
 export const ChartSubMenuHeads = ({ ids }) =>
   ids.length ? <ChartHeadsContainer ids={ids} /> : null
