+++ src/domains/charts/charts/index.js
