+++ src/domains/charts/index.js
@@ -7,4 +7,8 @@ export * from "./chartSubMenu"
 export * from "./menuGroup"
 export * from "./subMenu"
 
+export { default as getChartHeads } from "./getChartHeads"
+export { default as getChartMenu } from "./getChartMenu"
+export { default as getMenu } from "./getMenu"
+
 export { default as DashboardMenuProvider } from "./provider"
