+++ src/domains/charts/menuGroup/heads.js
@@ -1,6 +1,6 @@
 import React from "react"
-import { ChartHeadsContainer } from "domains/charts/providers/chartHead"
-import { withMenuGroup } from "domains/charts/providers/menuGroup"
+import { ChartHeadsContainer } from "domains/charts/chartHead"
+import { withMenuGroup } from "domains/charts/menuGroup"
 
 export const MenuGroupHeads = ({ ids }) => (ids.length ? <ChartHeadsContainer ids={ids} /> : null)
